+++ src/cljs/quil/sketch.cljs
@@ -22,19 +22,7 @@
     (u/resolve-constant-key mode rendering-modes)))
 
 (defn set-size [applet width height]
-  (when-let [el (.-quil-canvas applet)]
-    ; p5js creates a <canvas> element inside provided <div> element
-    ; we need to resize only the canvas as outer div will adapt automatically
-    (let [inner-canvas (.querySelector el "canvas")]
-      (.resizeCanvas applet width height)
-      (.setAttribute inner-canvas "width" width)
-      (.setAttribute inner-canvas "height" height)
-      (aset (.-style inner-canvas) "width" (str width "px"))
-      (aset (.-style inner-canvas) "height" (str height "px"))
-      (set! (.-width applet)
-            (.parseInt js/window (style/getComputedStyle inner-canvas "width")))
-      (set! (.-height applet)
-            (.parseInt js/window (style/getComputedStyle inner-canvas "height"))))))
+  (.resizeCanvas applet width height))
 
 (defn size
   ([width height]